02-support-ticket-rag/requirements.txt
sentence-transformers>=2.2.2
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
python-dotenv>=1.0.0
openai>=1.0.0
google-genai>=1.62.0
